[Feat] 삭제 예외처리

+++ src/main/java/outfoot/outfootserver/confirm/service/ConfirmService.java
@@ -87,6 +87,9 @@ public class ConfirmService {
         if (confirm.getImageUrl() != null) {
             fileUploader.deleteFile(confirm.getImageUrl(), "confirm");
         }
+        if (confirm.getMember() != member) {
+            throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
+        }
         confirmRepository.delete(confirm);
     }
